requirements.txt
streamlit
requests
python-dotenv
faster-whisper
moviepy
